requirements.txt
Flask==3.0.0
google-genai
python-dotenv==1.0.0
gunicorn==21.2.0
